requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0
django-pgcrypto-fields==2.6.0
rules==3.3
django-cors-headers==4.3.1
qrcode==7.4.2
httpx==0.27.0
huey==2.5.0
webauthn==2.0.0
#remove following lines after removing Djoser code, if ever
djangorestframework-simplejwt==5.3.0
django-templated-mail==1.1.1
